requirements.txt
fastapi
uvicorn
python-multipart
pillow
ultralytics
torch
gunicorn